requirements.txt
Requests
Python-dotenv